# typescript-angular-client-generated/model/apiResponse.ts
/**
 * Words
 * Words backend API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: example@example.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ApiResponse { 
    code?: number;
    type?: string;
    message?: string;
}