# typescript-angular-client-generated/model/word.ts
/**
 * Words
 * Words backend API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: example@example.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Word { 
    id?: number;
    name: string;
    photoUrls: Array<string>;
}